refactor: remove redundant casts (Rector, strictly safe subset)
Behaviour-preserving cleanup only, hand-picked rules (15 files):
- RecastingRemovalRector — drop casts already matching the value's type
- RemoveConcatAutocastRector — drop (string) casts redundant in concatenation
- RemoveUnusedNonEmptyArrayBeforeForeachRector — drop redundant if-guard before foreach

RemoveUnusedVariableAssignRector was evaluated and REJECTED: it dropped the
assignment of a security filter's return value in PageController
(`$relatedAlbums = $this->filterAlbumsByAccess(...)` -> bare call, which
discards the filtered result and would leak private/NSFW related albums), so
it is excluded. The broader Rector sets (readonly, switch->match, explicit-bool
compare, first-class-callable, constructor promotion) remain deferred to a
separately-reviewed pass.

Validated: PHPUnit 84 green; PHPStan delta zero; style re-checked with the
formatter.

+++ app/Controllers/Frontend/SearchController.php
@@ -56,7 +56,7 @@ final class SearchController extends BaseController
         }
 
         $total = (int) $result['total'];
-        $totalPages = (int) max(1, (int) ceil($total / $perPage));
+        $totalPages = max(1, (int) ceil($total / $perPage));
 
         return $this->view->render($response, 'frontend/search.twig', [
             'query' => $result['query'],
